+++ src/org/tn5250j/connectdialog/ConnectDialog.java
@@ -25,7 +25,6 @@
  */
 package org.tn5250j.connectdialog;
 
-import org.tn5250j.Configure;
 import org.tn5250j.ExternalProgramConfig;
 import org.tn5250j.TN5250jConstants;
 import org.tn5250j.gui.JSortTable;
